requirements.txt
flask==3.0.0
flask-cors==4.0.0
opencv-python==4.8.1.78
torch==2.1.2
torchvision==0.16.2
numpy==1.24.3
mediapipe==0.10.5
Pillow>=9.5.0
PyYAML>=6.0
requests>=2.31.0
scipy>=1.11.0
tqdm>=4.66.0
ultralytics==8.1.42
